setup image layout transitions
